Cleanup: move GpsManager into sensors package.
