+++ src/androidTest/java/de/dennisguse/opentracks/services/TrackRecordingServiceStateMachineTest.java
@@ -47,7 +47,7 @@ import java.util.concurrent.TimeoutException;
 
 import de.dennisguse.opentracks.data.ContentProviderUtils;
 import de.dennisguse.opentracks.data.models.Track;
-import de.dennisguse.opentracks.services.handlers.GpsStatusValue;
+import de.dennisguse.opentracks.sensors.GpsStatusValue;
 
 /**
  * Testing the states of TrackRecordingService.
